text2speech-azureAI-001/requirements.txt
azure-cognitiveservices-speech==1.34.0